pyproject.toml
[tool.poetry]
name = "games-api"
version = "0.1.0"
description = "API for creating Games"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
python-dotenv = "^0.19.0"
drf-yasg = "^1.20.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.1"
django-redis = "^5.0.0"
djangorestframework-simplejwt = "^4.7.2"
pytest = "^6.2.5"
pytest-django = "^4.4.0"
mixer = "^7.1.2"
celery = "^5.1.2"
django-celery-beat = "^2.2.1"
flower = "^1.0.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
